requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
kombu==5.4.2
oauthlib==3.2.2
packaging==24.2
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
